refactor(workspace): extract file tree model helpers
